+++ dnsimple/struct/__init__.py
@@ -13,6 +13,7 @@ from dnsimple.struct.domain_price import DomainPrice
 from dnsimple.struct.domain_registration import DomainRegistration, DomainRegistrationRequest
 from dnsimple.struct.domain_renewal import DomainRenewal, DomainRenewRequest
 from dnsimple.struct.domain_transfer import DomainTransfer, DomainTransferRequest
+from dnsimple.struct.domain_transfer_lock import DomainTransferLock
 from dnsimple.struct.domain_push import DomainPush
 from dnsimple.struct.email_forward import EmailForward, EmailForwardInput
 from dnsimple.struct.registrant import RegistrantChange, CheckRegistrantChangeInput, CreateRegistrantChangeInput, RegistrantChangeCheck