+++ dnsimple/struct/domain_transfer_lock.py
@@ -1,0 +1,9 @@
+from dataclasses import dataclass
+from dnsimple.struct.struct import Struct
+
+@dataclass
+class DomainTransferLock(Struct):
+    enabled = None
+
+    def __init__(self, data):
+        super().__init__(data)