+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
 FEATURES:
 
+- NEW: Added `get_domain_transfer_lock`, `enable_domain_transfer_lock`, and `disable_domain_transfer_lock` APIs to manage domain transfer locks. (dnsimple/dnsimple-python#435)
 - NEW: Added `list_registrant_changes`, `create_registrant_change`, `check_registrant_change`, `get_registrant_change`, and `delete_registrant_change` APIs to manage registrant changes. (dnsimple/dnsimple-python#433)
 
 ## 2.10.1